Add graph tests
